+++ grunt/config/less.js
@@ -1,10 +1,12 @@
+const constants = require('./../../lib/constants');
+
 module.exports = {
     default: {
         options: {
             compress: true
         },
         files: [{
-            dest: 'assets/session.min.css',
+            dest: `assets/${constants.assets.css}`,
             src: 'client/style.less'
         }]
     }