+++ RELEASES.md
@@ -1,5 +1,10 @@
 # Releases
 
+## 1.2.5 - 2018-01-09
+
+- Revert changes to `.getDomainByName()`.
+- Version named assets.
+
 ## 1.2.4 - 2018-01-08
 
 - `.getDomainByName(persistence, domainName)`.